restruct meaning
