+++ controllers/mathProblem/mathProblemGenerator.js
@@ -1,5 +1,5 @@
 const {generateScientificNotation} = require("./scientificNotation");
-const {generateMeaningOfExponents} = require("./meaningOfExponents");
+const {generateMeaningOfExponents} = require("./meaningOfExponents/meaningOfExponents");
 // const {generateOperationsOfExponents} = require("./operationsOfExponents")
 const {generateOperationsOfExponents} = require("./operationsOfExponents/operationsOfExponents");
